server/ml/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
xgboost==1.7.6
lightgbm==3.3.5
mlflow==2.5.0
psycopg2-binary==2.9.6
redis==4.6.0
sqlalchemy==2.0.19
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
ta==0.10.2
yfinance==0.2.28
ccxt==4.0.0
websocket-client==1.6.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
celery==5.3.1
